Add buffer and buffer window classes
A buffer represents a subset of the logcat data, holding only those
entries that have passed a set of filters. The filter part is not yet
implemented. In the future, the buffer will also manage user
interaction, such as jumping to earlier entries, automatic scrolling,
etc.

A buffer window is the curses part of a buffer.

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>