Add statusbar and statusbar window classes
The statusbar tracks information about the currently selected buffer,
such as its name and number of lines. In the future, the statusbar will
also accept user input (like ':' in less and vim).

The statusbar window is the curses part of a statusbar. The window
occupies the last line of the screen.

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>