Command/dispatch framework
An EVENT_COMMAND event contains an array of string which are passed on
to the new command framework and interpreted similarly to a shell
command. Commands do not have any return value; this may change in the
future.

Examples include 'quit', 'error "this is the error message"' and
'foobar --foo=1234'.

Internally, a command is implemented as a sub-class of _Command. The
base class manages an argparse parser; sub-classes are responsible for
preparing the parser with accepted arguments and for performing the
actual command. The same object is reused for all invocations of that
command. For this reason, the invocation call should not make any
changes to the object.

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>