Add command 'goto-line'
Add a command that makes the buffer jump to a given line. Surrounding
lines are displayed below, above or surrounding the given line (think
vim's zt, zb and zz commands).

Buffer now works in one of two modes: scrolling and non-scrolling. If
the buffer is in the first mode, it is expected to show the last logcat
entries and to automatically scroll to make room whenever a new entry is
added. Conversely, in non-scrolling mode, the buffer will not change
without user input.

A buffer starts out in scrolling mode.

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>